+++ .gitignore
@@ -975,7 +975,6 @@ packages/tools/update-readme-sponsors.js
 packages/tools/updateMarkdownDoc.js
 packages/tools/utils/discourse.js
 packages/tools/utils/loadSponsors.js
-packages/tools/utils/translation.test.js
 packages/tools/utils/translation.js
 packages/tools/website/build.js
 packages/tools/website/buildTranslations.js
